Revert "[FLINK-32223][runtime][security] Add Hive delegation token support"
This reverts commit 70e635983dc9aa4bd772042f34a04099256a8750.

It was decided to revert this issue due to compilation errors in hadoop_313 CI stage. See the Jira issue for further details.
